fix to separate tokens bar and toc
